LayoutTests/media/modern-media-controls/media-controller/ios/media-controller-stop-updates-in-fullscreen-expected.txt

Testing we stop updating controls while in fullscreen and then update them again when back to inline on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Obtained a 'play' event.

Entering fullscreen.
PASS media.webkitDisplayingFullscreen became true

Entered fullscreen
PASS media.paused is false
PASS window.internals.shadowRoot(media).querySelector('.media-controls').childElementCount became 0
PASS media.webkitDisplayingFullscreen became false

Exited fullscreen

PASS successfullyParsed is true

TEST COMPLETE
